Add files via upload
I wasn't able to git push from the terminal to include both folders. When I tried to git push and entered my credentials for the last step (getting them right from GitHub) the message back was failed and when I tried to upload the folders directly to GitHub it seemed to make my computer freeze. 

I needed help from TA's this week with the loops. I had my gameplay for iterating over rows then appending to lists. I wasn't able to nail the formatting of the txt printouts. Also my delimiter is off on the second one and I couldn't figure out how to replace it for the csvwriter